Extract edit button out of the component
